Remove `TokenScanner` dependency; move `getDirectTraits` logic to `AttributeFactory`
